Move ScoreWithRrcf and ModelWith* files to latis.ops.anomalies package
