+++ python/src/main/scala/latis/ops/anomalies/ScoreWithRrcf.scala
@@ -1,9 +1,7 @@
-package latis.ops
+package latis.ops.anomalies
 
 import jep.NDArray
 import jep.SharedInterpreter
-
-import latis.ops.anomalies.ModelTimeSeries
 
 /**
  * Defines an Operation that assigns anomaly scores to points in a univariate 
@@ -13,7 +11,7 @@ import latis.ops.anomalies.ModelTimeSeries
  * @param shingleSize the size of each shingle when shingling the time series
  * @param treeSize the size of each tree in the generated forest
  */
-case class ScoreWithRRCF(
+case class ScoreWithRrcf(
   numTrees: Int = 100,
   shingleSize: Int = 18,
   treeSize: Int = 256) extends ModelTimeSeries {
